fix build

+++ src/lib/inc/circular_queue.h
@@ -11,7 +11,7 @@
 #define cq_max_size(cq) (sizeof(cq.objects) / sizeof(cq.objects[0]))
 
 #define cq_is_full(cq) \
-    ((cq.rear + 1) % cq_max_size(cq) == cq.front)
+    (((cq.rear + 1) % (int) cq_max_size(cq)) == cq.front)
 
 #define cq_is_empty(cq) \
     (cq.front == -1)
@@ -22,40 +22,40 @@
         cq.front = cq.rear = -1;        \
     } while (0)
 
-#define cq_add(cq, data)                                        \
-    ({                                                          \
-        bool _ok;                                               \
-        spin_lock(&cq.lock);                                    \
-        _ok = !cq_is_full(cq);                                  \
-                                                                \
-        if (_ok) {                                              \
-            if ((cq.front) == -1) (cq.front) = 0;               \
-            (cq.rear) = ((cq.rear) + 1) % (cq_max_size(cq));    \
-            (cq.objects)[(cq.rear)] = (data);                   \
-            _ok = true;                                         \
-        }                                                       \
-                                                                \
-        spin_unlock(&cq.lock);                                  \
-        _ok;                                                    \
+#define cq_add(cq, data)                                              \
+    ({                                                                \
+        bool _ok;                                                     \
+        spin_lock(&cq.lock);                                          \
+        _ok = !cq_is_full(cq);                                        \
+                                                                      \
+        if (_ok) {                                                    \
+            if ((cq.front) == -1) (cq.front) = 0;                     \
+            (cq.rear) = ((cq.rear) + 1) % ((int)(cq_max_size(cq)));   \
+            (cq.objects)[(cq.rear)] = (data);                         \
+            _ok = true;                                               \
+        }                                                             \
+                                                                      \
+        spin_unlock(&cq.lock);                                        \
+        _ok;                                                          \
     })
 
-#define cq_pop(cq, data)                                        \
-    ({                                                          \
-        bool _ok;                                               \
-        spin_lock(&cq.lock);                                    \
-        _ok = !cq_is_empty(cq);                                 \
-                                                                \
-        if (_ok) {                                              \
-            *data = cq.objects[cq.front];                       \
-                                                                \
-            if (cq.front == cq.rear)                            \
-                cq.front = cq.rear = -1;                        \
-            else                                                \
-                cq.front = (cq.front + 1) % cq_max_size(cq);    \
-        }                                                       \
-        spin_unlock(&cq.lock);                                  \
-                                                                \
-        _ok;                                                    \
+#define cq_pop(cq, data)                                              \
+    ({                                                                \
+        bool _ok;                                                     \
+        spin_lock(&cq.lock);                                          \
+        _ok = !cq_is_empty(cq);                                       \
+                                                                      \
+        if (_ok) {                                                    \
+            *data = cq.objects[cq.front];                             \
+                                                                      \
+            if (cq.front == cq.rear)                                  \
+                cq.front = cq.rear = -1;                              \
+            else                                                      \
+                cq.front = (cq.front + 1) % ((int)cq_max_size(cq));   \
+        }                                                             \
+        spin_unlock(&cq.lock);                                        \
+                                                                      \
+        _ok;                                                          \
     })
 
 #endif //BAO_CIRCULAR_QUEUE_H
